pyproject.toml
[tool.poetry]
name = "tektome-api"
version = "0.1.0"
description = "Tektome assignment"
authors = ["Tanaphol SUEBCHUA"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
pytest-fastapi = "^0.1.0"
minio-async = "^1.0.1"
minio = "^7.2.5"
langchain = "^0.1.16"
python-dotenv = "^1.0.1"
jq = "^1.7.0"
tiktoken = "^0.6.0"
langchain-openai = "^0.1.3"
qdrant-client = "^1.8.2"
python-multipart = "^0.0.9"
requests = "^2.31.0"
pydantic-settings = "^2.2.1"
celery = "^5.4.0"
uvicorn = "^0.29.0"
redis = "^5.0.4"
black = "^24.4.1"
pytest = "^8.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
